Third week

+++ Server/src/main/java/com/service/UserService.java
@@ -2,6 +2,8 @@ package com.service;
 
 import com.entity.User;
 import com.repository.UserRepository;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -16,11 +18,19 @@ public class UserService {
             return false;
         }
         else {
-            userRepository.save(newUser);
-            return true;
+            try {
+                userRepository.save(newUser);
+                return true;
+            }
+            catch (DataIntegrityViolationException e){
+                return false;
+            }
         }
     }
     public void deleteUser(User user){
+        if(!userRepository.existsById(user.getName())){
+            throw new InvalidDataAccessApiUsageException("User not exist!");
+        }
         userRepository.delete(userRepository.getUserById(user.getName()));
     }
 }
